Revert "update for end-to-end-performance-test"
This reverts commit d51c390e180bf11ee1fb9751bb94525511be4433, reversing
changes made to 0dcd1e2f5c7d20a23bbb5699f4e0c6c32583daba.

+++ flink-runtime/src/main/java/org/apache/flink/runtime/operators/resettable/SpillingResettableMutableObjectIterator.java
@@ -53,9 +53,9 @@ public class SpillingResettableMutableObjectIterator<T> implements ResettableMut
 	
 	protected final TypeSerializer<T> serializer;
 	
-	private long elementCount;
+	private int elementCount;
 	
-	private long currentElementNum;
+	private int currentElementNum;
 	
 	protected final SpillingBuffer buffer;
